修改导航栏的字体的样式和配色

+++ Meizi/Classes/ZCMainDetailViewController.m
@@ -89,7 +89,7 @@
     
     [self.tableView reloadData];
     
-    NSLog(@"%ld", _dataSource.count);
+    NSLog(@"%lu", (unsigned long)_dataSource.count);
     
 }
